+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Bugfixes
 
  - Fix inability to change Input Period ([#393](https://github.com/kizniche/mycodo/issues/393))
+ - Fix Exception while reading the GPIO pin of Edge Conditional ([#387](https://github.com/kizniche/mycodo/issues/387))
 
 ### Miscellaneous